Moved status tests and pretests to new package layout
