+++ integration-tests/src/test/java/io/testerra/test/pretest_status/ExpectedFailsTests.java
@@ -19,18 +19,19 @@
  * under the License.
  */
 
-package eu.tsystems.mms.tic.testframework.test.execution;
+package io.testerra.test.pretest_status;
 
 import eu.tsystems.mms.tic.testframework.annotations.Fails;
 import eu.tsystems.mms.tic.testframework.annotations.Retry;
 import eu.tsystems.mms.tic.testframework.report.Status;
 import eu.tsystems.mms.tic.testframework.report.model.context.MethodContext;
 import eu.tsystems.mms.tic.testframework.testing.TesterraTest;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import org.testng.Assert;
-import org.testng.annotations.Test;
 
 public class ExpectedFailsTests extends TesterraTest implements TestStatusTest {
