+++ integration-tests/src/test/java/io/testerra/test/pretest_status/GenerateSimpleTestStatusTest.java
@@ -1,11 +1,9 @@
-package eu.tsystems.mms.tic.testframework.test.execution;
+package io.testerra.test.pretest_status;
 
-import eu.tsystems.mms.tic.testframework.annotations.InfoMethod;
 import eu.tsystems.mms.tic.testframework.execution.testng.AssertCollector;
 import eu.tsystems.mms.tic.testframework.execution.testng.OptionalAssert;
 import eu.tsystems.mms.tic.testframework.report.Status;
 import eu.tsystems.mms.tic.testframework.testing.TesterraTest;
-
 import org.testng.Assert;
 import org.testng.SkipException;
 import org.testng.annotations.Test;
@@ -48,6 +46,6 @@ public class GenerateSimpleTestStatusTest extends TesterraTest {
     public void testAssertCollector() {
         AssertCollector.fail("failed1");
         AssertCollector.fail("failed2");
-        AssertCollector.assertTrue(true,"passed1");
+        AssertCollector.assertTrue(true, "passed1");
     }
 }
