+++ integration-tests/src/test/java/io/testerra/test/pretest_status/RetryTests.java
@@ -19,25 +19,24 @@
  * under the License.
  */
 
-package eu.tsystems.mms.tic.testframework.test.execution;
+package io.testerra.test.pretest_status;
 
 import eu.tsystems.mms.tic.testframework.annotations.Fails;
 import eu.tsystems.mms.tic.testframework.annotations.NoRetry;
 import eu.tsystems.mms.tic.testframework.report.Status;
 import eu.tsystems.mms.tic.testframework.report.model.context.MethodContext;
 import eu.tsystems.mms.tic.testframework.testing.TesterraTest;
+import org.openqa.selenium.WebDriverException;
+import org.openqa.selenium.json.JsonException;
+import org.openqa.selenium.remote.UnreachableBrowserException;
+import org.testng.Assert;
+import org.testng.annotations.Test;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
-import org.openqa.selenium.WebDriverException;
-import org.openqa.selenium.json.JsonException;
-import org.openqa.selenium.remote.UnreachableBrowserException;
-import org.testng.Assert;
-import org.testng.annotations.Test;
 
 /**
  * Testclass to test the behaviour of TestNG in case of a retried test if it is a precondition of another test
