+++ integration-tests/src/test/resources/Pretest_StatusTests.xml
@@ -1,0 +1,14 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+<suite name="Execution Features" verbose="10" configfailurepolicy="continue" thread-count="10" parallel="false">
+
+    <test name="Parallel" parallel="methods">
+<!--        <groups>-->
+<!--            <run>-->
+<!--                <exclude name="SEQUENTIAL"/>-->
+<!--            </run>-->
+<!--        </groups>-->
+        <packages>
+            <package name="io.testerra.test.pretest_status"/>
+        </packages>
+    </test>
+</suite>
