+++ integration-tests/src/test/resources/log4j2-pretest.xml
@@ -3,7 +3,7 @@
     <Appenders>
         <File name="PreTestFile" fileName="logs/pre-test-log4j.log" append="false">
             <PatternLayout>
-                <Pattern>[%t][%p]%contextIds: %c{2} - %m%n</Pattern>
+                <Pattern>%d{dd.MM.yyyy HH:mm:ss.SSS} [%t][%p]%contextIds: %c{2} - %m%n</Pattern>
             </PatternLayout>
         </File>
         <Async name="Async">
